#272: Enable back button for search results
It is not possible to use the form implementation anymore, as the POST data is gone with the current request stack implementation.
Using GET doesn't solve this directly, as the query part is currenlty ignored by the current stack implementation, and adding it seemed to create more problems than solving it.
By redesigning the form with a custom handler that will convert it in a simple get request, full navigation is fixed without adjusting to much in the other logic.